Add request validations for all inputs

+++ app/Http/Controllers/DeleteController.php
@@ -29,6 +29,7 @@ class DeleteController extends Controller
     public function delete_stock_products(Request $request){
         $delete_ids = $request->except('_token');
         $entries_deleted = 0;
+        $entries_total = count($delete_ids);
         
         foreach($delete_ids as $id){
             /**
@@ -36,44 +37,63 @@ class DeleteController extends Controller
 
             //Stocks variables:
             $stocks_product = Stock::find($id);
-            $command_id = $stocks_product->command_id;
-            $stock_quantity = $stocks_product->quantity;
 
-            //Commands variables:
-            $commands_product = Command::find($command_id);
-            $command_quantity = $commands_product->quantity;
-            $alert_stock = $commands_product->alert_stock;
-            $newquantity = $command_quantity - $stock_quantity;
-            if($newquantity <= $alert_stock){
-                $must_buy = 1;
+            if(!empty($stocks_product)){
+                $command_id = $stocks_product->command_id;
+                $stock_quantity = $stocks_product->quantity;
+
+                //Commands variables:
+                $commands_product = Command::find($command_id);
+                $command_quantity = $commands_product->quantity;
+                $alert_stock = $commands_product->alert_stock;
+                $newquantity = $command_quantity - $stock_quantity;
+                if($newquantity <= $alert_stock){
+                    $must_buy = 1;
+                }
+                else {
+                    $must_buy = 0;
+                }
+
+                //Update commands:
+                $commands_product->update([
+                    'quantity' => $newquantity,
+                    'must_buy' => $must_buy,
+                ]);
+                
+                /** 
+                 * remove product from stocks */
+                $stocks_product->delete();
+
+                $entries_deleted += 1;                
             }
             else {
-                $must_buy = 0;
+                $entries_deleted += 0;
             }
-
-            //Update commands:
-            $commands_product->update([
-                'quantity' => $newquantity,
-                'must_buy' => $must_buy,
-            ]);
-            
-            /** 
-             * remove product from stocks */
-            $stocks_product->delete();
-
-            $entries_deleted += 1;
         }
 
-        if($entries_deleted == 1){
-            return redirect('stocks')->with('success', $entries_deleted . ' entry deleted !');
-        }
-        elseif($entries_deleted > 1){
-            return redirect('stocks')->with('success', $entries_deleted . ' entries deleted !');
+        if($entries_total != $entries_deleted){
+            $difference = $entries_total - $entries_deleted;
+            if($entries_deleted == 1){
+                return redirect('stocks')->with('success', "{$entries_deleted} entry deleted, {$difference} couldn't be deleted");
+            }
+            elseif($entries_deleted > 1){
+                return redirect('stocks')->with('success', "{$entries_deleted} entries deleted, {$difference} couldn't be deleted");
+            }
+            elseif($entries_deleted == 0) {
+                return redirect('stocks')->with('error', "There is an error, no entry deleted");
+            }
         }
         else {
-            return redirect('stocks')->with('error', 'There is an error, no entry deleted !');
+            if($entries_deleted == 1){
+                return redirect('stocks')->with('success', "{$entries_deleted} entry deleted !");
+            }
+            elseif($entries_deleted > 1){
+                return redirect('stocks')->with('success', "{$entries_deleted} entries deleted !");
+            }
+            elseif($entries_deleted == 0){
+                return redirect('stocks')->with('error', "There is an error, no entry deleted !");
+            }            
         }
-        
     }
 
     //Commands
@@ -97,30 +117,50 @@ class DeleteController extends Controller
     public function delete_command_products(Request $request){
         $delete_ids = $request->except('_token');
         $entries_deleted = 0;
+        $entries_total = count($delete_ids);
         
         foreach($delete_ids as $delete_id){
-            // remove product from stocks
-            $stocks_products = Stock::where('command_id', $delete_id);
-
-            $stocks_products->delete();
-
-            //remove entry from total stocks and all related entries in stocks
+            
             $commands_product = Command::find($delete_id);
 
-            $commands_product->delete();
-            
-            $entries_deleted += 1;
+            if(!empty($commands_product)){
+                // remove product from stocks
+                $stocks_products = Stock::where('command_id', $delete_id);
+
+                $stocks_products->delete();
+
+                //remove entry from total stocks
+                $commands_product->delete();
+                
+                $entries_deleted += 1;                
+            }
+            else {
+                $entries_deleted += 0;
+            }
         }
 
-        if($entries_deleted == 1){
-            return redirect('stocks')->with('success', $entries_deleted . ' entry deleted !');
-        }
-        elseif($entries_deleted > 1){
-            return redirect('stocks')->with('success', $entries_deleted . ' entries deleted !');
+        if($entries_total != $entries_deleted){
+            $difference = $entries_total - $entries_deleted;
+            if($entries_deleted == 1){
+                return redirect('total-stocks')->with('success', "{$entries_deleted} entry deleted, {$difference} couldn't be deleted");
+            }
+            elseif($entries_deleted > 1){
+                return redirect('total-stocks')->with('success', "{$entries_deleted} entries deleted, {$difference} couldn't be deleted");
+            }
+            elseif($entries_deleted == 0) {
+                return redirect('total-stocks')->with('error', "There is an error, no entry deleted");
+            }
         }
         else {
-            return redirect('stocks')->with('error', 'There is an error, no entry deleted !');
+            if($entries_deleted == 1){
+                return redirect('total-stocks')->with('success', "{$entries_deleted} entry deleted !");
+            }
+            elseif($entries_deleted > 1){
+                return redirect('total-stocks')->with('success', "{$entries_deleted} entries deleted !");
+            }
+            elseif($entries_deleted == 0){
+                return redirect('total-stocks')->with('error', "There is an error, no entry deleted !");
+            }            
         }
-        
     }
 }
